refactor(runtime): move resolveCliInterrupts to its own module; break cycle
resolveCliInterrupts moves into cliInterruptResolution.ts, sitting above the
shared core and runPolicyHandler, and now delegates to buildDecider +
resolveInterrupts. It passes NO policy to the endpoint (finding 1): run's policy
acts only in-chain, so a surfaced, policy-approved effect is decided afresh
(rejected non-interactively, prompted interactively) rather than auto-approved.

runPolicyHandler no longer imports from interrupts.ts, breaking the
interrupts <-> runPolicyHandler runtime cycle (lint:structure clean). index.ts
re-points the public resolveCliInterrupts export; generated code is unchanged.
